bookings añadidos, falta probar bruno

+++ app/Http/Controllers/UserController.php
@@ -3,49 +3,44 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UpdateUserRequest;
+use App\Http\Resources\BookingResource;
 use App\Http\Resources\UserResource;
+use Illuminate\Support\Facades\Auth;
 use App\Models\User;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
     public function index()
     {
-        /* if(Auth::user()->cannot('viewAny', User::class)) { */
-        /*     return response()->json([ */
-        /*         'message' => "You cannot perform this action" */
-        /*     ], 401); */
-        /* } */
+         if(Auth::user()->cannot('viewAny', User::class)) {
+             return response()->json([
+                 'message' => "You cannot perform this action"
+             ], 401);
+         }
 
         return UserResource::collection(User::all());
     }
 
-    /**
-     * Display the specified resource.
-     */
+
     public function show(User $user)
     {
-        /* if(Auth::user()->cannot('view', $user)) { */
-        /*     return response()->json([ */
-        /*         'message' => "You cannot perform this action" */
-        /*     ], 401); */
-        /* } */
+         if(Auth::user()->cannot('view', $user)) {
+             return response()->json([
+                 'message' => "You cannot perform this action"
+             ], 401);
+         }
 
         return new UserResource($user);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(UpdateUserRequest $request, User $user)
     {
-        /* if(Auth::user()->cannot('update', $user)) { */
-        /*     return response()->json([ */
-        /*         'message' => "You cannot perform this action" */
-        /*     ], 401); */
-        /* } */
+         if(Auth::user()->cannot('update', $user)) {
+             return response()->json([
+                 'message' => "You cannot perform this action"
+             ], 401);
+         }
 
         $user->update($request->validated());
 
@@ -54,16 +49,13 @@ class UserController extends Controller
         ]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(User $user)
     {
-        /* if(Auth::user()->cannot('delete', $user)) { */
-        /*     return response()->json([ */
-        /*         'message' => "You cannot perform this action" */
-        /*     ], 401); */
-        /* } */
+         if(Auth::user()->cannot('delete', $user)) {
+             return response()->json([
+                'message' => "You cannot perform this action"
+             ], 401);
+         }
 
         $user->delete();
 
@@ -72,10 +64,8 @@ class UserController extends Controller
         ]);
     }
 
-    public function bookings(User $user) {
-        /*
-        * TODO: Falta crear el recurso de bookins y devolver una coleccion de los bookings
-        * del usuario haciendo uso de la relacion bookings
-        */
+    public function bookings(User $user)
+    {
+        return BookingResource::collection($user->bookings);
     }
 }
